+++ src/project.rs
@@ -161,10 +161,13 @@ impl FuzzProject {
 
         let mut rustflags: String = "-Cpasses=sancov-module \
                                      -Cllvm-args=-sanitizer-coverage-level=4 \
-                                     -Cllvm-args=-sanitizer-coverage-trace-compares \
                                      -Cllvm-args=-sanitizer-coverage-inline-8bit-counters \
                                      -Cllvm-args=-sanitizer-coverage-pc-table"
             .to_owned();
+
+        if !build.no_trace_compares {
+            rustflags.push_str(" -Cllvm-args=-sanitizer-coverage-trace-compares");
+        }
 
         if !build.no_cfg_fuzzing {
             rustflags.push_str(" --cfg fuzzing");